src/requirements.txt
tensorflow[and-cuda]==2.19.0
numpy
pillow
scipy